requirements.txt
streamlit
tensorflow-cpu
numpy
nltk
pickle-mixin
